minor refactoring for better code structure
