+++ smart-contracts/contracts/pro-vault/src/msg.rs
@@ -1,15 +1,15 @@
 use cosmwasm_schema::cw_serde;
 use cw_vault_standard::{VaultStandardExecuteMsg, VaultStandardQueryMsg};
 use cosmwasm_std::Uint128;
-// use crate::vault::{provault, config};
 use crate::vault::provault::VaultRunningState;
 use crate::vault::config::Config;
-use crate::strategy::strategy::{Strategy, StrategyKey};
-use serde::{Serialize,Deserialize};
+use crate::strategy::strategy::{Strategy, StrategyKey, StrategyAction}; 
+use serde::{Serialize, Deserialize};
+use crate::vault::query::VaultQueryMsg; 
+
 use schemars::JsonSchema;
 
-// TODO - Complex instantiation support to be added so provault can be fully orchastreted 
-// using an elegant json file.
+// Pro vault instantiate message structure
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct InstantiateMsg {
     pub thesis: String,      // The general thesis of the vault
@@ -18,12 +18,11 @@ pub struct InstantiateMsg {
 }
 
 
+// Pro vault query message enums types.
 #[cw_serde]
 pub enum QueryMsg {
     GetAllStrategies {},
-    GetVaultConfig {},    
-    GetVaultRunningState {},
-
+    VaultQuery(VaultQueryMsg), // Use VaultQueryMsg for vault-related queries
 }
 
 #[cw_serde]
@@ -32,49 +31,21 @@ pub struct MigrateMsg {}
 #[cw_serde]
 pub enum ProExtensionExecuteMsg {
     MyVariant1 {
-        /// The amount of base tokens to deposit.
         amount: Uint128,
-        /// The optional recipient of the vault token. If not set, the caller
-        /// address will be used instead.
         recipient: Option<String>,
     },
-
-    // TODO - Combine all vault related enums as Vault Actions or VaultAdmin Action.
     UpdateRunningState {
         new_state: VaultRunningState,
-        // Placeholder for running state details
     },
-
-    UpdateVaultOwner {
-        // Placeholder for vault owner details
+    UpdateVaultOwner {},
+    UpdateStrategyOwner {},
+    CreateStrategy {
+        name: String,
+        description: String,
     },
-
-
-    UpdateStrategyOwner {
-        // Placeholder for strategy owner details
+    ExecStrategyActions {
+        action: StrategyAction,
     },
-
-    // TODO - Adding adaptors, configuring adaptors, adding Strategy Control Owner, Adaptor Control Owner
-    CreateStrategy { name: String, description: String },  
-
-
-
-    ExecStrategyActions { 
-        action : StrategyAction,
-    },
-}
-
-#[cw_serde]
-pub enum StrategyAction {
-    DistributeFundWithPresetAdaptorRatio, // Distributing funds across adaptors as per preset ratios
-    DistributeFundWithCustomAdaptorRatios { custom_ratios: String }, // CustomAdaptorRatio (A1:R1, A2:R2, A3:R3)
-    RemoveAdaptor { adaptor: String }, // Remove Adaptor Ai
-    AddNewAdaptor { adaptor: String }, // Add a new adaptor of type Ai. Should fail if already one is present of type A1.
-    UpdateStrategyParams ,
-    //{ // Placeholder for updating strategy parameters // e.g., update ratio, remove adaptor, enable/disable strategy or adaptor
-    //},
-    UpdateAdaptorRunningState { adaptor: String },
-    UpdateStrategyRunningState,
 }
 
 #[cw_serde]
@@ -86,20 +57,4 @@ pub enum ExtensionExecuteMsg {
 pub type ExecuteMsg = VaultStandardExecuteMsg<ExtensionExecuteMsg>;
 
 
-// QUERY RESPONSES 
-// TODO - Structure to code to the right module and files .
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
-pub struct StrategyInfoResponse {
-    pub strategies: Vec<Strategy>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct VaultConfigResponse {
-    pub config: Config,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct VaultRunningStateResponse {
-    pub state: VaultRunningState,
-    pub last_statechange_bh: u64,
-}
+
