+++ smart-contracts/contracts/pro-vault/src/vault/mod.rs
@@ -1,2 +1,3 @@
 pub mod config;
 pub mod provault;
+pub mod query;
